Refactor gatsby-node with new main query
Added a new main query in gatsby node which is in charge
of providing the data to the blog post listing page with
pagination and also to the category and tags pages with
pagination also. This commit only adds working blog post
pagination as it was before, next step is extending
pagination to category and tags pages.